Add files via upload
To get started with the Desktop News Notifier, we require two libraries: feedparser and notify2.Feedparser will parse the feed that we will get from the URL. We will use notify2 for the desktop notification purpose. Other than these two libararies, we will use OS and time lib. Once you are done with the installation import both libraries in the program.